Added multiple features of Linkis-UDF version

+++ linkis-computation-governance/linkis-manager/label-manager/src/main/scala/org/apache/linkis/manager/label/service/impl/DefaultNodeLabelRemoveService.scala
@@ -24,9 +24,10 @@ import org.apache.linkis.manager.label.entity.em.EMInstanceLabel
 import org.apache.linkis.manager.label.entity.engine.EngineInstanceLabel
 import org.apache.linkis.manager.label.service.{NodeLabelRemoveService, NodeLabelService}
 import org.apache.linkis.manager.persistence.LabelManagerPersistence
-import org.apache.linkis.message.annotation.Receiver
+import org.apache.linkis.rpc.serializer.annotation.Receiver
 import org.apache.linkis.protocol.label.NodeLabelRemoveRequest
 import org.apache.linkis.manager.label.conf.LabelCommonConfig
+import org.apache.linkis.manager.label.exception.LabelRuntimeException
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.stereotype.Service
 
@@ -45,6 +46,9 @@ class DefaultNodeLabelRemoveService extends NodeLabelRemoveService with Logging 
   @Receiver
   override def removeNodeLabel(nodeLabelRemoveRequest: NodeLabelRemoveRequest): Unit = {
     info(s"Start to remove labels from node ${nodeLabelRemoveRequest.getServiceInstance}")
+    if(nodeLabelRemoveRequest.getServiceInstance == null) {
+      throw new LabelRuntimeException(130001, "ServiceInstance in request is null, please check label remove request!")
+    }
     nodeLabelService.removeLabelsFromNode(nodeLabelRemoveRequest.getServiceInstance)
     val persistenceLabel = if (nodeLabelRemoveRequest.isEngine) {
       val engineLabel = labelFactory.createLabel(classOf[EngineInstanceLabel])
